api/pyproject.toml
[tool.poetry]
name = "home-api"
version = "0.1.0"
description = ""
authors = ["Sasha Chistyh <sashchystyh@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
pre-commit = "^2.20.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
python-dotenv = "^0.21.0"
aerich = "^0.7.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
fastapi-utils = "^0.2.1"
strawberry-graphql = {extras = ["fastapi"], version = "^0.151.2"}
mangum = "^0.17.0"



[tool.aerich]
tortoise_orm = "core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
